requirements.txt
pandas==2.0.2
matplotlib==3.7.1
openpyxl==3.1.2
ipykernel==6.23.1
PyYAML==6.0
scipy==1.10.1
scikit-learn==1.2.2